+++ src/main/java/gp/project/nodes/ProgramNode.java
@@ -16,7 +16,7 @@ public class ProgramNode extends Node {
         super(tree, another);
     }
 
-    public double run(List<Double> input) {
+    public int run(List<Integer> input) {
         // TODO (in the future)
 
         return input.get(input.size() - 1);
